Convert review rating back to an integer

+++ database/migrations/2018_04_26_141254_create_reviews_table.php
@@ -17,8 +17,8 @@ class CreateReviewsTable extends Migration
             $table->increments('id');
             //$table->integer('restaurant_id');
             $table->text('comment');
-            //$table->integer('rating');
-            $table->float('rating', 2, 1);
+            $table->integer('rating');
+            //$table->float('rating', 2, 1);
             $table->timestamps();
         });
     }
